Se corrigió la implementación y persistencia de las clases Categoría y Tags

+++ backend/src/main/java/com/nocountry/cms/models/Tag.java
@@ -16,7 +16,7 @@ public class Tag {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id_tag;
+    private Long id_tag;
     private String nombre;
 
     @ManyToMany(mappedBy = "tags")
